moved noflo specific code into its own file
getting error `runtime.runtime.on is not a function`, although the same
error can be observed in commit 59f7a21d85aba087fc0bf9836bd7c3ffa9853ae5
